Added the ability to create a .ct file (string-based data file)
A .ct or CockTail file is created upon a user entering strings into the
create scene and pressing the ADD RECIPE button. Current logic for this
function is still very primitive and lacks the ability for text to wrap
around in the text-fields.

Currently functions are implemented in the controller Class. Eventually,
these will have to move to a model, through which, we call functions.

Next, I will prioritize development of the ability to read .ct files in
the View.fxml scene.